Fix updateStatus to use validating icon
The updateStatus function is called immediately on a field even before getting a response (such as with 'remote'), and uses the "invalid" feedbackIcons rather than the "validating" icons. The logic clearly states that it should show a valid icon if there are no "not validated" (pre validated) errors/messages, and to show an invalid icon otherwise. Unfortunately, for 'remote' validators that may take time to respond, this will immediately display the 'invalid' icon in the meantime rather than the 'validating' icon.